# docs/swagger.yaml
basePath: /api
definitions:
...
  bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel:
    properties:
      code: {}
      message:
        example: error
        type: string
      status:
        example: error
        type: string
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.FileOut:
    properties:
      file:
        type: string
      kind:
        type: string
      status:
        type: string
    type: object
...
host: localhost:9567
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a go fp transaction api docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GO FP TRANSACTION API DUCUMENTATION
  version: "1.0"
paths:
...
  /v1/files/upload:
    post:
      consumes:
      - application/json
      description: Upload transaction from CSV file
      produces:
      - application/json
      responses:
        "200":
          description: Response indicates that the request succeeded and the resources
            has been fetched and transmitted in the message body
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.FileOut'
        "400":
          description: Bad request error. This can happen if there is an error while
            create account
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "500":
          description: Internal server error. This can happen if there is an error
            while create account
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Upload transaction file
      tags:
      - Files
...
  /v2/files/upload:
    post:
      consumes:
      - application/json
      description: Upload transaction from CSV file
      produces:
      - application/json
      responses:
        "200":
          description: Response indicates that the request succeeded and the resources
            has been fetched and transmitted in the message body
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.FileOut'
        "400":
          description: Bad request error. This can happen if there is an error while
            create account
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "500":
          description: Internal server error. This can happen if there is an error
            while create account
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Upload transaction file
      tags:
      - Files
...
schemes:
- http
swagger: "2.0"